backend/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.13.0

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0

# HTTP client
httpx>=0.26.0

# Media analysis
pymediainfo>=6.0.0

# Plex integration
plexapi>=4.15.0

# Data export
pandas>=2.0.0

# Development
python-dotenv>=1.0.0
